remove gcc warnings

+++ src/Core/VipFunctional.h
@@ -894,7 +894,7 @@ public:
 	/// @brief Returns all functions that match exactly given argument types.
 	function_list_type exactMatch(const VipTypeList& lst) const
 	{
-		if (lst.size() > NArgs)
+		if (lst.size() > static_cast<int>(NArgs))
 			return function_list_type();
 		function_list_type res;
 		for (const function_type& f : m_functions) {
